requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pandas==2.2.2
python-dateutil==2.9.0.post0
numpy==2.1.0
pyyaml==6.0.2
ortools==9.10.4067
pytest==8.3.2
